Make EntityList components singular
We generally use singular names for components
(e.g. RecordingMerge), so there's no reason to have
these as plural, which also looks weird.
