+++ root/area/artists.tt
@@ -2,7 +2,7 @@
     <h2>[% l('Artists') %]</h2>
 
     [% IF artists.size %]
-      [% React.embed(c, 'components/ArtistsList', { artists => artists, showBeginEnd => 1, showRatings => 1})
+      [% React.embed(c, 'components/ArtistList', { artists => artists, showBeginEnd => 1, showRatings => 1})
            WRAPPER 'components/with-pager.tt' %]
     [% ELSE %]
       [% l('This area is not currently associated with any artists.') %]
