+++ root/components/ReleaseGroupList.js
@@ -22,7 +22,7 @@ import parseDate from '../static/scripts/common/utility/parseDate';
 import RatingStars from './RatingStars';
 import SortableTableHeader from './SortableTableHeader';
 
-type ReleaseGroupsListHeaderProps = {|
+type ReleaseGroupListHeaderProps = {|
   ...SeriesItemNumbersRoleT,
   +$c: CatalystContextT,
   +checkboxes?: string,
@@ -32,7 +32,7 @@ type ReleaseGroupsListHeaderProps = {|
   +sortable?: boolean,
 |};
 
-type ReleaseGroupsListEntryProps = {|
+type ReleaseGroupListEntryProps = {|
   ...SeriesItemNumbersRoleT,
   +$c: CatalystContextT,
   +checkboxes?: string,
@@ -42,7 +42,7 @@ type ReleaseGroupsListEntryProps = {|
   +showRatings?: boolean,
 |};
 
-type ReleaseGroupsListProps = {|
+type ReleaseGroupListProps = {|
   ...SeriesItemNumbersRoleT,
   +$c: CatalystContextT,
   +checkboxes?: string,
@@ -53,7 +53,7 @@ type ReleaseGroupsListProps = {|
   +sortable?: boolean,
 |};
 
-const ReleaseGroupsListHeader = ({
+const ReleaseGroupListHeader = ({
   $c,
   checkboxes,
   groupByType,
@@ -61,7 +61,7 @@ const ReleaseGroupsListHeader = ({
   seriesItemNumbers,
   showRatings,
   sortable,
-}: ReleaseGroupsListHeaderProps) => (
+}: ReleaseGroupListHeaderProps) => (
   <thead>
     <tr>
       {$c.user_exists && checkboxes ? (
@@ -112,7 +112,7 @@ const ReleaseGroupsListHeader = ({
   </thead>
 );
 
-const ReleaseGroupsListEntry = ({
+const ReleaseGroupListEntry = ({
   $c,
   checkboxes,
   index,
@@ -120,7 +120,7 @@ const ReleaseGroupsListEntry = ({
   releaseGroup,
   seriesItemNumbers,
   showRatings,
-}: ReleaseGroupsListEntryProps) => (
+}: ReleaseGroupListEntryProps) => (
   <tr className={loopParity(index)} key={releaseGroup.id}>
     {$c.user_exists && checkboxes ? (
       <td>
@@ -166,7 +166,7 @@ const ReleaseGroupsListEntry = ({
   </tr>
 );
 
-const ReleaseGroupsListTable = ({
+const ReleaseGroupListTable = ({
   $c,
   checkboxes,
   groupByType,
@@ -175,9 +175,9 @@ const ReleaseGroupsListTable = ({
   seriesItemNumbers,
   showRatings,
   sortable,
-}: ReleaseGroupsListProps) => (
+}: ReleaseGroupListProps) => (
   <table className="tbl release-group-list">
-    <ReleaseGroupsListHeader
+    <ReleaseGroupListHeader
       $c={$c}
       checkboxes={checkboxes}
       groupByType={groupByType}
@@ -188,7 +188,7 @@ const ReleaseGroupsListTable = ({
     />
     <tbody>
       {releaseGroups.map((releaseGroup, index) => (
-        <ReleaseGroupsListEntry
+        <ReleaseGroupListEntry
           $c={$c}
           checkboxes={checkboxes}
           groupByType={groupByType}
@@ -203,7 +203,7 @@ const ReleaseGroupsListTable = ({
   </table>
 );
 
-const ReleaseGroupsList = ({
+const ReleaseGroupList = ({
   $c,
   checkboxes,
   groupByType,
@@ -212,7 +212,7 @@ const ReleaseGroupsList = ({
   seriesItemNumbers,
   showRatings,
   sortable,
-}: ReleaseGroupsListProps) => {
+}: ReleaseGroupListProps) => {
   const groupedReleaseGroups = groupBy(releaseGroups, 'typeName');
   return (
     groupByType ? (
@@ -226,7 +226,7 @@ const ReleaseGroupsList = ({
                 : releaseGroupType(releaseGroupsOfType[0])
               }
             </h3>
-            <ReleaseGroupsListTable
+            <ReleaseGroupListTable
               $c={$c}
               checkboxes={checkboxes}
               groupByType
@@ -241,7 +241,7 @@ const ReleaseGroupsList = ({
       })
     ) : (
       // TODO: When converting usages to React, please check MBS-10155.
-      <ReleaseGroupsListTable
+      <ReleaseGroupListTable
         $c={$c}
         checkboxes={checkboxes}
         order={order}
@@ -254,4 +254,4 @@ const ReleaseGroupsList = ({
   );
 };
 
-export default withCatalystContext(ReleaseGroupsList);
+export default withCatalystContext(ReleaseGroupList);
