+++ root/edit/details/merge_instruments.tt
@@ -2,13 +2,13 @@
   <tr>
     <th>[% l('Merge:') %]</th>
     <td>
-      [%~ React.embed(c, 'components/InstrumentsList', { instruments => edit.display_data.old }) ~%]
+      [%~ React.embed(c, 'components/InstrumentList', { instruments => edit.display_data.old }) ~%]
     </td>
   </tr>
   <tr>
     <th>[% l('Into:') %]</th>
     <td>
-      [%~ React.embed(c, 'components/InstrumentsList', { instruments => [edit.display_data.new] }) ~%]
+      [%~ React.embed(c, 'components/InstrumentList', { instruments => [edit.display_data.new] }) ~%]
     </td>
   </tr>
 </table>
