+++ root/edit/details/merge_release_groups.tt
@@ -2,13 +2,13 @@
     <tr>
         <th>[% l('Merge:') %]</th>
         <td>
-            [% React.embed(c, 'components/ReleaseGroupsList', { releaseGroups => edit.display_data.old }) # TODO: On conversion to React, please check MBS-10155. %]
+            [% React.embed(c, 'components/ReleaseGroupList', { releaseGroups => edit.display_data.old }) # TODO: On conversion to React, please check MBS-10155. %]
         </td>
     </tr>
     <tr>
         <th>[% l('Into:') %]</th>
         <td>
-            [% React.embed(c, 'components/ReleaseGroupsList', { releaseGroups => [edit.display_data.new] }) # TODO: On conversion to React, please check MBS-10155. %]
+            [% React.embed(c, 'components/ReleaseGroupList', { releaseGroups => [edit.display_data.new] }) # TODO: On conversion to React, please check MBS-10155. %]
         </td>
     </tr>
 </table>
