+++ root/instrument/recordings.tt
@@ -4,7 +4,7 @@
     [%- IF recordings.size -%]
         <form action="[% c.uri_for_action('/recording/merge_queue') %]"
             method="post">
-            [% React.embed(c, 'components/RecordingsList', { recordings => recordings, checkboxes => 'add-to-merge', showRatings => 1, showInstrumentCredits => 1, instrumentCredits => instrument_credits })
+            [% React.embed(c, 'components/RecordingList', { recordings => recordings, checkboxes => 'add-to-merge', showRatings => 1, showInstrumentCredits => 1, instrumentCredits => instrument_credits })
                 WRAPPER 'components/with-pager.tt' -%]
             [% form_submit(l('Add selected recordings for merging')) WRAPPER form_row IF c.user_exists %]
         </form>
