+++ root/label/index.tt
@@ -15,7 +15,7 @@
     [%- IF releases.size -%]
       <form action="[% c.uri_for_action('/release/merge_queue') %]"
             method="post">
-        [% React.embed(c, 'components/ReleasesList', { releases => releases, checkboxes => 'add-to-merge', filterLabel => label})
+        [% React.embed(c, 'components/ReleaseList', { releases => releases, checkboxes => 'add-to-merge', filterLabel => label})
             WRAPPER 'components/with-pager.tt' -%]
         [% form_submit(l('Add selected releases for merging')) WRAPPER form_row IF c.user_exists %]
       </form>
