+++ root/series/index.tt
@@ -15,32 +15,32 @@
       IF series.type.item_entity_type == 'event';
         '<h2>' _ l('Event') _ '</h2>';
         IF entities.size;
-          React.embed(c, 'components/EventsList', { events => entities, seriesItemNumbers => series_item_numbers, showArtists => 1, showLocation => 1, showRatings => 1 })
+          React.embed(c, 'components/EventList', { events => entities, seriesItemNumbers => series_item_numbers, showArtists => 1, showLocation => 1, showRatings => 1 })
           WRAPPER 'components/with-pager.tt';
         END;
       ELSIF series.type.item_entity_type == 'recording';
         '<h2>' _ l('Recordings') _ '</h2>';
         IF entities.size;
-          React.embed(c, 'components/RecordingsList', { recordings => entities, showRatings => 1, seriesItemNumbers => series_item_numbers})
+          React.embed(c, 'components/RecordingList', { recordings => entities, showRatings => 1, seriesItemNumbers => series_item_numbers})
           WRAPPER 'components/with-pager.tt';
         END;
       ELSIF series.type.item_entity_type == 'release';
         '<h2>' _ l('Releases') _ '</h2>';
         IF entities.size;
-          React.embed(c, 'components/ReleasesList', { releases => entities, seriesItemNumbers => series_item_numbers})
+          React.embed(c, 'components/ReleaseList', { releases => entities, seriesItemNumbers => series_item_numbers})
           WRAPPER 'components/with-pager.tt';
         END;
       ELSIF series.type.item_entity_type == 'release_group';
         '<h2>' _ l('Release Groups') _ '</h2>';
         IF entities.size;
-          React.embed(c, 'components/ReleaseGroupsList',
+          React.embed(c, 'components/ReleaseGroupList',
                   { releaseGroups => entities, seriesItemNumbers = series_item_numbers, showRatings => 1  }) # TODO: On conversion to React, please check MBS-10155.
           WRAPPER 'components/with-pager.tt';
         END;
       ELSIF series.type.item_entity_type == 'work';
         '<h2>' _ l('Works') _ '</h2>';
         IF entities.size;
-          React.embed(c, 'components/WorksList', { works => entities, seriesItemNumbers = series_item_numbers, showRatings => 1})
+          React.embed(c, 'components/WorkList', { works => entities, seriesItemNumbers = series_item_numbers, showRatings => 1})
           WRAPPER 'components/with-pager.tt';
         END;
       END
