+++ root/server/components.js
@@ -211,20 +211,20 @@ module.exports = {
   'artist/ArtistHeader': require('../artist/ArtistHeader'),
   'collection/CollectionHeader': require('../collection/CollectionHeader'),
   'components/Aliases': require('../components/Aliases'),
-  'components/AreasList': require('../components/AreasList'),
-  'components/ArtistsList': require('../components/ArtistsList'),
-  'components/EventsList': require('../components/EventsList'),
+  'components/AreaList': require('../components/AreaList'),
+  'components/ArtistList': require('../components/ArtistList'),
+  'components/EventList': require('../components/EventList'),
   'components/ExpirationTime': require('../components/ExpirationTime'),
-  'components/InstrumentsList': require('../components/InstrumentsList'),
-  'components/LabelsList': require('../components/LabelsList'),
-  'components/PlacesList': require('../components/PlacesList'),
-  'components/RecordingsList': require('../components/RecordingsList'),
+  'components/InstrumentList': require('../components/InstrumentList'),
+  'components/LabelList': require('../components/LabelList'),
+  'components/PlaceList': require('../components/PlaceList'),
+  'components/RecordingList': require('../components/RecordingList'),
   'components/Relationships': require('../components/Relationships'),
-  'components/ReleaseGroupsList': require('../components/ReleaseGroupsList'),
-  'components/ReleasesList': require('../components/ReleasesList'),
+  'components/ReleaseGroupList': require('../components/ReleaseGroupList'),
+  'components/ReleaseList': require('../components/ReleaseList'),
   'components/SeriesList': require('../components/SeriesList'),
   'components/UserAccountTabs': require('../components/UserAccountTabs'),
-  'components/WorksList': require('../components/WorksList'),
+  'components/WorkList': require('../components/WorkList'),
   'edit/components/EditSidebar': require('../edit/components/EditSidebar'),
   'edit/components/EditSummary': require('../edit/components/EditSummary'),
   'edit/components/Vote': require('../edit/components/Vote'),
